Move footer to designated folder
