+++ src/layout/Layout.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Footer from '~components/Footer'
+import Footer from '~components/footer/Footer'
 import NavBar from '~components/navbar/NavBar'
 import SEO from '~components/SEO'
